+++ core/components/colorpicker/elements/tv/output/options/colorpicker.php
@@ -4,10 +4,9 @@
  *
  * @package colorpicker
  * @subpackage outputoptions_render
- *
- * @var modX $modx
  */
 
+/** @var modX $modx */
 $corePath = $modx->getOption('colorpicker.core_path', null, $modx->getOption('core_path') . 'components/colorpicker/');
 
 return $modx->smarty->fetch($corePath . 'elements/tv/output/tpl/colorpicker.options.tpl');