+++ source/js/mgr/colorpickerfield.js
@@ -11,76 +11,49 @@
  */
 
 Ext.namespace("Ext.ux", "Ext.ux.menu", "Ext.ux.form");
-if (Ext.version.substr(0, 1) === "2") {
-    Ext.ux.menu.ColorItem = function (config) {
-        if (!config) config = {};
-        config.style = "width:400px;";
-        Ext.ux.menu.ColorItem.superclass.constructor.call(this, new Ext.ux.ColorPicker(config), config);
-        this.picker = this.component;
-        this.addEvents('select');
-        this.picker.on("render", function (picker) {
-            picker.getEl().swallowEvent("click");
+
+Ext.ux.menu.ColorMenu = Ext.extend(Ext.menu.Menu, {
+    enableScrolling: false,
+    hideOnClick: true,
+    initComponent: function () {
+        Ext.apply(this, {
+            plain: true,
+            showSeparator: false,
+            items: this.picker = new Ext.ux.ColorPicker(Ext.apply({
+                style: 'width:400px;'
+            }, this.initialConfig))
         });
-        this.picker.on("select", this.onSelect, this);
-    };
-    Ext.extend(Ext.ux.menu.ColorItem, Ext.menu.Adapter, {
-        // private
-        onSelect: function (picker, color) {
-            this.fireEvent("select", this, color, picker);
-            Ext.ux.menu.ColorItem.superclass.handleClick.call(this);
+        this.picker.purgeListeners();
+        Ext.ux.menu.ColorMenu.superclass.initComponent.call(this);
+        this.relayEvents(this.picker, ['select']);
+        this.on('select', this.menuHide, this);
+        if (this.handler) {
+            this.on('select', this.handler, this.scope || this)
         }
-    });
-    Ext.ux.menu.ColorMenu = function (config) {
-        Ext.ux.menu.ColorMenu.superclass.constructor.call(this, config);
-        this.plain = true;
-        var ci = new Ext.ux.menu.ColorItem(config);
-        this.add(ci);
-        this.picker = ci.picker;
-        this.relayEvents(ci, ["select"]);
-    };
-    Ext.extend(Ext.ux.menu.ColorMenu, Ext.menu.Menu, {
-        beforeDestroy: function () {
-            this.picker.destroy();
+    },
+    menuHide: function () {
+        if (this.hideOnClick) {
+            this.hide(true);
         }
-    });
-} else {
-    Ext.ux.menu.ColorMenu = Ext.extend(Ext.menu.Menu, {
-        enableScrolling: false,
-        hideOnClick: true,
-        initComponent: function () {
-            Ext.apply(this, {
-                plain: true,
-                showSeparator: false,
-                items: this.picker = new Ext.ux.ColorPicker(Ext.apply({
-                    style: 'width:400px;'
-                }, this.initialConfig))
-            });
-            this.picker.purgeListeners();
-            Ext.ux.menu.ColorMenu.superclass.initComponent.call(this);
-            this.relayEvents(this.picker, ['select']);
-            this.on('select', this.menuHide, this);
-            if (this.handler) {
-                this.on('select', this.handler, this.scope || this)
-            }
-        },
-        menuHide: function () {
-            if (this.hideOnClick) {
-                this.hide(true);
-            }
-        }
-    });
-}
+    }
+});
+
 Ext.ux.form.ColorPickerField = Ext.extend(Ext.form.TriggerField, {
     triggerClass: 'x-form-colorpicker-trigger',
     initComponent: function () {
         Ext.ux.form.ColorPickerField.superclass.initComponent.call(this);
-        this.menu = new Ext.ux.menu.ColorMenu();
+        console.log(this.id);
+var trigger = this;
+        document.addEventListener("DOMContentLoaded",function() {
+            this.picker = new ColorPicker(trigger.id, this.getValue());
+        });
+        //this.menu = new Ext.ux.menu.ColorMenu();
     },
     setValue: function (v) {
         if (/^[0-9a-fA-F]{6}$/.test(v)) {
             Ext.ux.form.ColorPickerField.superclass.setValue.apply(this, arguments);
-            var i = this.menu.picker.rgbToHex(this.menu.picker.invert(this.menu.picker.hexToRgb(v)));
-            this.el.applyStyles('background: #' + v + '; color: #' + i + ';');
+            //var i = this.menu.picker.rgbToHex(this.menu.picker.invert(this.menu.picker.hexToRgb(v)));
+            this.el.applyStyles('background: #' + v + '; color: #fff' + ';');
         }
     },
     onDestroy: function () {
@@ -96,8 +69,8 @@ Ext.ux.form.ColorPickerField = Ext.extend(Ext.form.TriggerField, {
         Ext.ux.form.ColorPickerField.superclass.onBlur.call(this);
         var v = this.getValue();
         if (/^[0-9a-fA-F]{6}$/.test(v)) {
-            var i = this.menu.picker.rgbToHex(this.menu.picker.invert(this.menu.picker.hexToRgb(v)));
-            this.el.applyStyles('background: #' + v + '; color: #' + i + ';');
+            //var i = this.menu.picker.rgbToHex(this.menu.picker.invert(this.menu.picker.hexToRgb(v)));
+            this.el.applyStyles('background: #' + v + '; color: #fff' + ';');
         } else this.el.applyStyles('background: #ffffff; color: #000000;');
     },
     menuListeners: {