+++ core/components/colorpicker/docs/changelog.md
@@ -5,11 +5,21 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [1.0.5] - 2020-07-23
+## [2.0.0] - 2022-02-05
+
+### Added
+
+- Code refactoring
+- Full MODX 3 compatibility
+- Colorpicker on base of mdbassit/Coloris
+- Define swatches in the input properties of the template variable
+- Optional restrict the colorpicker to swatches
+- Allow alpha values in the template variable input and output
+- RGB and HSL template variable values
 
 ### Changed
 
-- Fix an allowBlank issue
+- The hex template variable (default) database values are prefixed with a "#" character during the update
 
 ## [1.0.4] - 2019-08-17